+++ BambooTracker/configuration.hpp
@@ -183,7 +183,7 @@ public:
 		EditInstrument, FollowMode, DuplicateOrder, ClonePatterns, CloneOrder, ReplaceInstrument,
 		ExpandPattern, ShrinkPattern, FineDecreaseValues, FineIncreaseValues, CoarseDecreaseValues,
 		CoarseIncreaseValuse, ExpandEffect, ShrinkEffect, PrevHighlighted, NextHighlighted,
-		IncreasePatternSize, DecreasePatternSize
+		IncreasePatternSize, DecreasePatternSize, IncreaseEditStep, DecreaseEditStep
 	};
 	void setShortcuts(std::unordered_map<ShortcutAction, std::string> shortcuts);
 	std::unordered_map<ShortcutAction, std::string> getShortcuts() const;