+++ BambooTracker/gui/mainwindow.hpp
@@ -177,6 +177,7 @@ private:
 	std::unique_ptr<QShortcut> focusPtnSc_, focusOdrSc_, focusInstSc_, instAddSc_;
 	std::unique_ptr<QAction> playAndStopSc_, playStepSc_, goPrevOdrSc_, goNextOdrSc_;
 	std::unique_ptr<QAction> prevInstSc_, nextInstSc_, incPtnSizeSc_, decPtnSizeSc_;
+	std::unique_ptr<QAction> incEditStepSc_, decEditStepSc_;
 	void setShortcuts();
 
 	// Dialogs