+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 ## Added
 - Shortcut settings
-- 2 new shortcuts
+- 4 new shortcuts
   - Increase/Decrease pattern size
+  - Increase/Decrease edit step
 
 ## Changed
 - [#198] - Raise the maximum of tick rate to 511Hz (thanks [@djmaximum17])